requirements.txt
python-dotenv
slack-bolt
pyepics
numpy
